Estudos

+++ Curso_em_Video_mundo_I_II_III/I/017.py
@@ -1,4 +1,4 @@
 pro = float(input('qual é o preço do produto? '))
-des = float(pro - (pro * 20 / 100))
+des = pro - (pro * 20 / 100)
 print('o pruduto que custava {} vai sair em pormoção com desconto de 20% por R${:.3}'.format(pro, des))
